Fixed Issue #86, Fixed rendering of Coal Generator's burn down

+++ src/main/java/com/theundertaker11/geneticsreborn/blocks/coalgenerator/GRTileEntityCoalGenerator.java
@@ -214,7 +214,7 @@ public class GRTileEntityCoalGenerator extends TileEntity implements ITickable {
     }
 
     public double percentage() {
-        return (double) (this.burnTime / this.currentItemBurnTime) * 100;
+        return (double) ((double)this.burnTime / (double)this.currentItemBurnTime) * 100;
     }
 
     public static boolean isEnergyTE(TileEntity te) {
